backend/requirements.txt
anthropic>=0.49.0
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic[email]==2.12.5
pydantic-settings==2.6.1
supabase==2.28.2
PyJWT==2.10.1
slowapi==0.1.9
python-multipart==0.0.17
httpx==0.28.1
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-asyncio==0.24.0
reportlab>=4.0.0
